+++ L4D2AddonAssistant.Core/AddonNode.cs
@@ -32,6 +32,8 @@ namespace L4D2AddonAssistant
 
         private long? _fileSize = null;
 
+        private DateTime _creationTime = DateTime.Now;
+
         public AddonNode(AddonRoot root, AddonGroup? group = null)
         {
             OnInitSelf();
@@ -199,6 +201,12 @@ namespace L4D2AddonAssistant
         {
             get => _fileSize;
             private set => NotifyAndSetIfChanged(ref _fileSize, value);
+        }
+
+        public DateTime CreationTime
+        {
+            get => _creationTime;
+            set => NotifyAndSetIfChanged(ref _creationTime, value);
         }
 
         internal CancellationToken DestructionCancellationToken => _destructionCancellationTokenSource.Token;
@@ -559,18 +567,24 @@ namespace L4D2AddonAssistant
 
             if (RequireFile)
             {
+                var fullFilePath = FullFilePath;
+
                 try
                 {
-                    var path = FullFilePath;
-                    if (!File.Exists(path) && !Directory.Exists(path))
+                    if (!File.Exists(fullFilePath) && !Directory.Exists(fullFilePath))
                     {
                         AddProblem(new AddonFileNotExistProblem(this));
                     }
                 }
                 catch (Exception ex)
                 {
-                    Log.Error(ex, "Exception occurred during AddonNode.OnPostCheck");
+                    LogException(ex);
                 }
+            }
+
+            void LogException(Exception ex)
+            {
+                Log.Error(ex, "Exception occurred during AddonNode.OnPostCheck");
             }
         }
 
@@ -597,12 +611,17 @@ namespace L4D2AddonAssistant
         {
             save.IsEnabled = IsEnabled;
             save.Name = Name;
+            save.CreationTime = CreationTime;
         }
 
         protected virtual void OnLoadSave(AddonNodeSave save)
         {
             IsEnabled = save.IsEnabled;
             Name = save.Name;
+            if (save.CreationTime != default)
+            {
+                CreationTime = save.CreationTime;
+            }
         }
 
         internal virtual void OnInitSelf()