+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- add: creation time display
+- add: sort by creation time
 # v0.2.0
 - improve: AddonNodeView
 - fix: WorkshopVpkAddon.FullVpkFilePath is sometimes not updated